Add extra file exists check to fix an error

+++ runner.js
@@ -70,9 +70,12 @@ function runContract() {
       if (fs.existsSync("syntest/statistics/statistics.csv")) {
         fs.copyFileSync("syntest/statistics/statistics.csv", `results/${contract}/statistics.csv`);
       }
-      const testFiles = fs.readdirSync("syntest/tests");
-      for (let file of testFiles) {
-        fs.copyFileSync(`syntest/tests/${file}`, `results/${contract}/tests/${file}`);
+
+      if (fs.existsSync("syntest/tests")) {
+        const testFiles = fs.readdirSync("syntest/tests");
+        for (let file of testFiles) {
+          fs.copyFileSync(`syntest/tests/${file}`, `results/${contract}/tests/${file}`);
+        }
       }
     }
